Add subscription metadata to message interface.
Insufficient subscription metadata is passed in the standard
metadata ``AMQP::Headers`` structure about what routing key,
exchange name, etc. were used for the specification of the
message's target.  This will be included now, which enables
a route selector, similar to what I have done with the Mothership,
to be used as a single AMQP entrypoint.